All fields are optional for draft applications + more fixes

+++ src/students/index.php
@@ -28,25 +28,28 @@ $application =
         ], true
     );
 
+if (HTTP::post('submit') && $application->status == 'draft') {
+    $application->status = 'submitted';
+}
+
 $form = new Form($application);
-$isUploadRequired = (!$application->attachment || $application->status != 'draft');
+$isUploadRequired = (!$application->attachment && $application->status != 'draft');
 $file = new FileUpload('attachment', $isUploadRequired, ALLOWED_UPLOAD_EXTENSIONS, 10485760);
 
-// TODO: If file is invalid show error even if old file exists
 // TODO: Show error if email send fails
-if (HTTP::isPost() && $application->isValid() && ($file->isValid() || $application->attachment)) {
+if (HTTP::isPost() && $application->isValid() && $file->isValid()) {
     DB::beginTransaction();
 
     try {
-        if ($application->status == 'draft') {
-            $application->status = 'submitted';
-        }
-
         $oldFilename = $application->attachment;
         $today = date('Y-m-d');
         $newFilename = "$today-{$user->id}.{$file->extension()}"; // 2020-12-30-00100001.pdf
-        $file->move(UPLOAD_DIR, $newFilename);
-        $application->attachment = $newFilename;
+
+        if ($file->isUploaded()) {
+            $file->move(UPLOAD_DIR, $newFilename);
+            $application->attachment = $newFilename;
+        }
+
         $application->save();
 
         if ($application->status == 'submitted') {
@@ -55,18 +58,21 @@ if (HTTP::isPost() && $application->isValid() && ($file->isValid() || $applicati
                 $application->advisorEmail,
                 'CSTEM Scholars Grant Application Needs Review',
                 HTML::template('emails/application_submitted_advisor.php', $form)
-            )->send();
+            );
 
             // Email the student
             Mail::send(
                 $application->email,
                 'CSTEM Scholars Grant Application Submitted',
                 HTML::template('emails/application_submitted_student.php', $form)
-            )->send();
+            );
         }
 
         DB::commit();
-        FileUpload::delete(UPLOAD_DIR, $oldFilename);
+
+        if ($file->isUploaded() && $oldFilename != $newFilename) {
+            FileUpload::delete(UPLOAD_DIR, $oldFilename);
+        }
 
         if ($application->status != 'draft') {
             HTTP::redirect('thank_you.php');
